backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asgiref>=3.9.1",
    "celery>=5.5.3",
    "cryptography>=45.0.5",
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-celery-beat>=2.8.1",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "email-validator>=2.2.0",
    "g4f>=0.6.2.8",
    "groq>=0.31.1",
    "nodriver>=0.47.0",
    "platformdirs>=4.4.0",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "sqlparse>=0.5.3",
]
